+++ OpenGLFramework/GameController.h
@@ -31,7 +31,7 @@ public:
 	virtual void Initialize();
 	virtual void ProcessInput(const GLint& Key, const GLint& Action, const GLint& Mode) = 0;
 	virtual void ProcessMouseMove(const GLdouble& dX, const GLdouble& dY) = 0;
-	virtual void Update(const GLdouble& dt) = 0;
+	virtual void Update(const GLfloat& dt) = 0;
 	virtual void Render() = 0;
 
 protected:
